Remove padding from base64_encode output
base64_encode () may contain one or two = symbols in the output for padding. These are not needed when doing base64_decode() but when fed to the symfony process component, it'll try to escape those special characters using escapeshellarg(). This PHP function has a limit and produces the following error:

"Argument exceeds the allowed length of 131072 bytes"

Simplest solution is to just trim the =s.